Removed Multichain Signature type
