+++ crypto-shared/Cargo.toml
@@ -9,7 +9,7 @@ serde = "1"
 borsh = "1.3.0"
 near-account-id = "1"
 serde_json = "1"
-near-sdk = {version = "=5.0.0-alpha.1", features = ["unstable"]}
+near-sdk = {version = "=5.1.0", features = ["unstable"]}
 
 # TODO ENSURE THIS ISN'T BUILT INTO THE NODE
 [target.'cfg(target_arch = "wasm32")'.dependencies]
