+++ crypto-shared/src/kdf.rs
@@ -1,14 +1,12 @@
-use crate::types::{PublicKey, ScalarExt};
+use crate::types::{PublicKey, ScalarExt, SignatureResponse};
 use anyhow::Context;
 use k256::{
     ecdsa::{RecoveryId, Signature, VerifyingKey},
     elliptic_curve::{point::AffineCoordinates, sec1::ToEncodedPoint, CurveArithmetic},
-    pkcs8::DecodePublicKey,
     sha2::{Digest, Sha256},
-    AffinePoint, Scalar, Secp256k1,
+    Scalar, Secp256k1,
 };
 use near_account_id::AccountId;
-use near_sdk::env;
 
 // Constant prefix that ensures epsilon derivation values are used specifically for
 // near-mpc-recovery with key derivation protocol vX.Y.Z.
@@ -34,20 +32,13 @@ pub fn derive_key(public_key: PublicKey, epsilon: Scalar) -> PublicKey {
     (<Secp256k1 as CurveArithmetic>::ProjectivePoint::GENERATOR * epsilon + public_key).to_affine()
 }
 
-#[derive(Debug)]
-pub struct MultichainSignature {
-    pub big_r: AffinePoint,
-    pub s: Scalar,
-    pub recovery_id: u8,
-}
-
 // try to get the correct recovery id for this signature by brute force.
 pub fn into_eth_sig(
     public_key: &k256::AffinePoint,
     big_r: &k256::AffinePoint,
     s: &k256::Scalar,
     msg_hash: Scalar,
-) -> anyhow::Result<MultichainSignature> {
+) -> anyhow::Result<SignatureResponse> {
     let public_key = public_key.to_encoded_point(false);
     let signature = k256::ecdsa::Signature::from_scalars(x_coordinate(big_r), s)
         .context("cannot create signature from cait_sith signature")?;
@@ -59,11 +50,7 @@ pub fn into_eth_sig(
     .context("unable to use 0 as recovery_id to recover public key")?
     .to_encoded_point(false);
     if public_key == pk0 {
-        return Ok(MultichainSignature {
-            big_r: *big_r,
-            s: *s,
-            recovery_id: 0,
-        });
+        return Ok(SignatureResponse::new(*big_r, *s, 0));
     }
 
     let pk1 = VerifyingKey::recover_from_prehash(
@@ -74,11 +61,7 @@ pub fn into_eth_sig(
     .context("unable to use 1 as recovery_id to recover public key")?
     .to_encoded_point(false);
     if public_key == pk1 {
-        return Ok(MultichainSignature {
-            big_r: *big_r,
-            s: *s,
-            recovery_id: 1,
-        });
+        return Ok(SignatureResponse::new(*big_r, *s, 1));
     }
 
     anyhow::bail!("cannot use either recovery id (0 or 1) to recover pubic key")
