+++ integration-tests/chain-signatures/tests/actions/mod.rs
@@ -3,9 +3,7 @@ pub mod wait_for;
 use crate::MultichainTestContext;
 
 use cait_sith::FullSignature;
-use mpc_contract::SignatureResponse;
-use crypto_shared::SerializableScalar;
-use crypto_shared::{derive_epsilon, derive_key, into_eth_sig};
+use crypto_shared::{derive_epsilon, derive_key, into_eth_sig, SerializableScalar, SignatureResponse};
 use elliptic_curve::sec1::ToEncodedPoint;
 use k256::ecdsa::VerifyingKey;
 use k256::elliptic_curve::ops::{Invert, Reduce};
